Fix rEnhancer for compatibility
